Update pest grammar
The program and expr_list become silent rules as they will not be referenced in the parser. The rule `WHITESPACE` is introduced to support whitespaces between types (atoms and invocations), otherwise the parser can only recognize one basic type. The atomic operation `@` must be present in `int` and `float` rule, to parse "+ 1" as name "+" followed by int "1" rather than int "+ 1".